HashtagProject/requirements.txt
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
pandas==2.2.0
pytest==8.0.1
pytz==2024.1
requests==2.31.0
